refactor(scanner)!: Always parse ScanCode license expressions
Remove the feature to toggle parsing of ScanCode license expressions.
Starting with ScanCode output format verion 3, license expressions are
first class citizens in ScanCode results, making this feature cumbersome
to maintain. It anyway is better implemented as a general post-processing
of scan results from all scanners that decomposes the licenses if
required.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>