helper-cli: Add a command to generate a flat analyzer result
The command defines a data model called `SimpleDefinitionFile` which is
read as input using any of the known file extensions. It outputs an
analyzer result with a single project containing the provided
dependencies as a flat list.

When a project does not use a package manager, currently the only way of
scanning it with ORT is to craft SPDX files and running the analyzer
against them. This command tries to provide a minimalistic alternative
to fullfil the requirements of the underlying need.

Note: This is a first experimental version. Also not that `VcsInfo` is
      deliberately not used, to 1. avoid nesting to in turn make it
      easier to generate the input file and to 2. make all attributes
      optional.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>